lib--mmar-carl/pyproject.toml
[project]
name = "mmar-carl"
# dynamic version is not supported yet on uv_build
version = "0.2.0"
description = "Collaborative Agent Reasoning Library with multi-step type support"
authors = [{name = "glazkov", email = "glazkov@airi.net"}, {name = "shaposhnikov", email = "shaposhnikov@airi.net"}, {name = "tagin", email = "tagin@airi.net"}]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.12"
keywords = []
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.11.7",
    "simpleeval>=1.0.0",
    "mmar-utils~=1.1.18",
    "openai>=1.0.0",
]

[project.optional-dependencies]
vector-search = [
    "faiss-cpu>=1.7.0",
    "fastembed>=0.7.3",
    "numpy>=1.21.0",
]
mcp = ["mcp>=1.0.0"]
langfuse = ["langfuse>=3.0.0,<4.0.0"]
pandas = ["pandas>=1.0.0"]
all = [
    "mmar-carl[vector-search]",
    "mmar-carl[mcp]",
    "mmar-carl[langfuse]",
    "mmar-carl[pandas]",
]

[dependency-groups]
ci = [
    "ruff>=0.4",
    "pytest>=8.2",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "mmar_carl"
source-exclude = [".ruff_cache"]

[tool.uv]
default-groups = ["ci"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    # External library warnings (faiss/swig) - not fixable in this codebase
    "ignore:builtin type .* has no __module__:DeprecationWarning",
]

[pytest]
python_files = [ "test_*.py" ]
testpaths = [ "tests" ]

[mypy]
ignore_missing_imports = true
exclude = "tests/fixtures/"
warn_unused_ignores = true
show_error_codes = true
